+++ library/Emphloyer/Scheduler/Backend.php
@@ -17,6 +17,25 @@ interface Backend {
   public function clear();
 
   /**
+   * List the entire schedule.
+   * @return \Iterator
+   */
+  public function allEntries();
+
+  /**
+   * Find a specific entry in the schedule using its id and return its attributes.
+   * @param mixed $id
+   * @return array|null
+   */
+  public function find($id);
+
+  /**
+   * Delete an entry from the schedule using its id
+   * @param mixed $id
+   */
+  public function delete($id);
+
+  /**
    * Schedule a job.
    * @param array $job Job to schedule
    * @param int $minute Minute to schedule on
@@ -24,7 +43,7 @@ interface Backend {
    * @param int $dayOfMonth Day of the month to schedule on
    * @param int $month Month to schedule on
    * @param int $dayOfWeek Week day to schedule on
-   * @return array Attributes of scheduled job
+   * @return array Attributes of scheduled entry
    */
   public function schedule(array $job, $minute = null, $hour = null, $dayOfMonth = null, $month = null, $dayOfWeek = null);